+++ zebra-node-services/src/mempool.rs
@@ -4,6 +4,8 @@
 
 use std::collections::HashSet;
 
+use tower::BoxError;
+
 use zebra_chain::transaction::{UnminedTx, UnminedTxId};
 
 /// A mempool service request.
@@ -86,4 +88,35 @@ impl From<UnminedTx> for Gossip {
     fn from(tx: UnminedTx) -> Self {
         Gossip::Tx(tx)
     }
+}
+
+/// A response to a mempool service request.
+///
+/// Responses can read the current set of mempool transactions,
+/// check the queued status of transactions to be downloaded and verified, or
+/// confirm that the mempool has been checked for newly verified transactions.
+#[derive(Debug)]
+pub enum Response {
+    /// Returns all transaction IDs from the mempool.
+    TransactionIds(HashSet<UnminedTxId>),
+
+    /// Returns matching transactions from the mempool.
+    ///
+    /// Since the [`TransactionsById`] request is unique,
+    /// the response transactions are also unique.
+    Transactions(Vec<UnminedTx>),
+
+    /// Returns matching cached rejected transaction IDs from the mempool,
+    RejectedTransactionIds(HashSet<UnminedTxId>),
+
+    /// Returns a list of queue results.
+    ///
+    /// These are the results of the initial queue checks.
+    /// The transaction may also fail download or verification later.
+    ///
+    /// Each result matches the request at the corresponding vector index.
+    Queued(Vec<Result<(), BoxError>>),
+
+    /// Confirms that the mempool has checked for recently verified transactions.
+    CheckedForVerifiedTransactions,
 }
